requirements.txt
langgraph==0.0.15
langchain==0.1.0
langchain-openai==0.0.2
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
apify-client==1.4.0
selenium==4.15.2
webdriver-manager==4.0.1 